renamed request.py to connection.py
