+++ starlite/types.py
@@ -31,10 +31,10 @@ except ImportError:  # pragma: no cover
     from typing import _GenericAlias as GenericAlias  # type: ignore
 
 if TYPE_CHECKING:  # pragma: no cover
+    from starlite.connection import Request
     from starlite.controller import Controller
     from starlite.datastructures import State
     from starlite.handlers import BaseRouteHandler
-    from starlite.request import Request
     from starlite.routing import Router
 else:
     Request = Any
