+++ tests/middleware/test_authentication.py
@@ -9,9 +9,9 @@ from starlette.status import (
 from starlette.websockets import WebSocketDisconnect
 
 from starlite import Starlite, create_test_client, get, websocket
+from starlite.connection import Request, WebSocket
 from starlite.exceptions import PermissionDeniedException
 from starlite.middleware import AbstractAuthenticationMiddleware, AuthenticationResult
-from starlite.request import Request, WebSocket
 
 
 class User(BaseModel):
